requirements.txt
pandas~=1.4.2
numpy~=1.22.3
mysql-connector-python~=8.0.29
beautifulsoup4~=4.11.1
lxml~=4.8.0